requirements.txt
torch
torchvision
torchaudio
ultralytics
fastapi
uvicorn
pillow
numpy
requests
python-multipart
jupyter
pydantic
python-dotenv
tqdm
datasets
roboflow
opencv-python
pyyaml
efficientnet-pytorch
matplotlib
pandas
